% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageApp-package.R
\docType{package}
\name{packageApp-package}
\alias{packageApp}
\alias{packageApp-package}
\title{packageApp: Shiny App as a Package}
\description{
Demonstrate `shiny` app as R package and test deployment on Posit Connect.
}
\author{
\strong{Maintainer}: Aleksander Chlebowski \email{aleksander.chlebowski@sanofi.com}

Authors:
\itemize{
  \item Aleksander Chlebowski \email{aleksander.chlebowski@sanofi.com}
}

}
\keyword{internal}
